オープンデータを入手: Rename components/OpenDataLink.vue to components/index/_shared/OpenDataLink.vue
