+++ components/index/CardsMonitoring/ConfirmedCasesDetails/Card.vue
@@ -62,9 +62,9 @@ import dayjs from 'dayjs'
 
 import AppLink from '@/components/_shared/AppLink.vue'
 import DataView from '@/components/DataView.vue'
+import OpenDataLink from '@/components/index/_shared/OpenDataLink.vue'
 // table タグとの衝突を避けるため ConfirmedCasesDetailsTable とする
 import ConfirmedCasesDetailsTable from '@/components/index/CardsMonitoring/ConfirmedCasesDetails/Table.vue'
-import OpenDataLink from '@/components/OpenDataLink.vue'
 import Data from '@/data/data.json'
 import formatConfirmedCases from '@/utils/formatConfirmedCases'
