+++ components/index/CardsReference/ConfirmedCasesByMunicipalities/Table.vue
@@ -33,7 +33,7 @@ import Vue from 'vue'
 
 import DataView from '@/components/DataView.vue'
 import DataViewDataSetPanel from '@/components/DataViewDataSetPanel.vue'
-import OpenDataLink from '@/components/OpenDataLink.vue'
+import OpenDataLink from '@/components/index/_shared/OpenDataLink.vue'
 
 export default Vue.extend({
   components: { DataView, DataViewDataSetPanel, OpenDataLink },
